// apps/admin-web/src/api/model/signInNewAgreementResponseDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Admin API
 * The Admin API description
 * OpenAPI spec version: 1.0
 */
import type { SignInNewAgreementResponseDtoTermType } from './signInNewAgreementResponseDtoTermType';

export interface SignInNewAgreementResponseDto {
  /** 약관 ID */
  id: number;
  /** 약관 내용 */
  content: string;
  /** 약관 버전 */
  version: string;
  /** 약관 유형 (필수/선택) */
  termType: SignInNewAgreementResponseDtoTermType;
  /** 약관 제목 */
  title: string;
  /** 약관 시작일 */
  startDate: string;
  /** 약관 종료일 */
  endDate: string;
  /**
   * 동합 동의 ID (미동의 시 null)
   * @nullable
   */
  agreementId?: number | null;
  /**
   * 동의 일시 (미동의 시 null)
   * @nullable
   */
  agreedAt?: string | null;
}